+++ crates/telio-lana/src/event_log_file.rs
@@ -70,41 +70,6 @@ fn get_logfile() -> std::io::Result<File> {
 pub mod moose {
     use serde::{Deserialize, Serialize};
 
-    /// Mock of moose::MeshnetappContext.
-    #[derive(Clone, Serialize, Deserialize)]
-    pub struct MeshnetappContext {
-        #[serde(rename = "application")]
-        pub application: MeshnetappContextApplication,
-    }
-
-    /// Mock of moose::MeshnetappContextApplicationConfig.
-    #[derive(Clone, Serialize, Deserialize)]
-    pub struct MeshnetappContextApplicationConfig {
-        #[serde(rename = "current_state")]
-        pub current_state: MeshnetappContextApplicationConfigCurrentState,
-    }
-
-    /// Mock of moose::MeshnetappContextApplicationConfigCurrentState.
-    #[derive(Clone, Serialize, Deserialize)]
-    pub struct MeshnetappContextApplicationConfigCurrentState {
-        #[serde(rename = "internal_meshnet")]
-        pub internal_meshnet: MeshnetappContextApplicationConfigCurrentStateInternalMeshnet,
-    }
-
-    /// Mock of moose::MeshnetappContextApplicationConfigCurrentStateInternalMeshnet.
-    #[derive(Clone, Serialize, Deserialize)]
-    pub struct MeshnetappContextApplicationConfigCurrentStateInternalMeshnet {
-        #[serde(rename = "fp")]
-        pub fp: Option<String>,
-    }
-
-    /// Mock of moose::MeshnetappContextApplication.
-    #[derive(Clone, Serialize, Deserialize)]
-    pub struct MeshnetappContextApplication {
-        #[serde(rename = "config")]
-        pub config: MeshnetappContextApplicationConfig,
-    }
-
     /// Mock of moose::LibtelioappLogLevel.
     #[derive(Debug, Clone, Deserialize, PartialEq, Serialize)]
     /// value of 'log_level'
@@ -254,12 +219,12 @@ pub mod moose {
     }
 
     #[allow(non_snake_case)]
-    pub fn set_context_application_libtelioapp_config_meshnetEnabled(
+    pub fn set_context_application_libtelioapp_config_currentState_meshnetEnabled(
         val: bool,
     ) -> std::result::Result<Result, Error> {
         let val = val.to_string();
         match super::event_log(
-            "set_context_application_libtelioapp_config_meshnetEnabled",
+            "set_context_application_libtelioapp_config_currentState_meshnetEnabled",
             Some(vec![val.as_str()]),
         ) {
             Ok(_) => Ok(Result::Success),
@@ -269,11 +234,11 @@ pub mod moose {
 
     #[allow(non_snake_case)]
     /// Mocked moose function.
-    pub fn set_context_application_libtelioapp_config_internalMeshnet_fpNat(
+    pub fn set_context_application_libtelioapp_config_currentState_internalMeshnet_fpNat(
         val: String,
     ) -> std::result::Result<Result, Error> {
         match super::event_log(
-            "set_context_application_libtelioapp_config_internalMeshnet_fpNat",
+            "set_context_application_libtelioapp_config_currentState_internalMeshnet_fpNat",
             Some(vec![val.as_str()]),
         ) {
             Ok(_) => Ok(Result::Success),
@@ -283,11 +248,11 @@ pub mod moose {
 
     #[allow(non_snake_case)]
     /// Mocked moose function.
-    pub fn set_context_application_libtelioapp_config_internalMeshnet_membersNat(
+    pub fn set_context_application_libtelioapp_config_currentState_internalMeshnet_membersNat(
         val: String,
     ) -> std::result::Result<Result, Error> {
         match super::event_log(
-            "set_context_application_libtelioapp_config_internalMeshnet_membersNat",
+            "set_context_application_libtelioapp_config_currentState_internalMeshnet_membersNat",
             Some(vec![val.as_str()]),
         ) {
             Ok(_) => Ok(Result::Success),
@@ -297,11 +262,11 @@ pub mod moose {
 
     #[allow(non_snake_case)]
     /// Mocked moose function.
-    pub fn set_context_application_libtelioapp_config_externalLinks(
+    pub fn set_context_application_libtelioapp_config_currentState_externalLinks(
         val: String,
     ) -> std::result::Result<Result, Error> {
         match super::event_log(
-            "set_context_application_libtelioapp_config_externalLinks",
+            "set_context_application_libtelioapp_config_currentState_externalLinks",
             Some(vec![val.as_str()]),
         ) {
             Ok(_) => Ok(Result::Success),
@@ -311,11 +276,11 @@ pub mod moose {
 
     #[allow(non_snake_case)]
     /// Mocked moose function.
-    pub fn set_context_application_libtelioapp_config_internalMeshnet_fp(
+    pub fn set_context_application_libtelioapp_config_currentState_internalMeshnet_fp(
         val: String,
     ) -> std::result::Result<Result, Error> {
         match super::event_log(
-            "set_context_application_libtelioapp_config_internalMeshnet_fp",
+            "set_context_application_libtelioapp_config_currentState_internalMeshnet_fp",
             Some(vec![val.as_str()]),
         ) {
             Ok(_) => Ok(Result::Success),
@@ -325,11 +290,11 @@ pub mod moose {
 
     #[allow(non_snake_case)]
     /// Mocked moose function.
-    pub fn set_context_application_libtelioapp_config_internalMeshnet_members(
+    pub fn set_context_application_libtelioapp_config_currentState_internalMeshnet_members(
         val: String,
     ) -> std::result::Result<Result, Error> {
         match super::event_log(
-            "set_context_application_libtelioapp_config_internalMeshnet_members",
+            "set_context_application_libtelioapp_config_currentState_internalMeshnet_members",
             Some(vec![val.as_str()]),
         ) {
             Ok(_) => Ok(Result::Success),
@@ -339,11 +304,11 @@ pub mod moose {
 
     #[allow(non_snake_case)]
     /// Mocked moose function.
-    pub fn set_context_application_libtelioapp_config_internalMeshnet_connectivityMatrix(
+    pub fn set_context_application_libtelioapp_config_currentState_internalMeshnet_connectivityMatrix(
         val: String,
     ) -> std::result::Result<Result, Error> {
         match super::event_log(
-            "set_context_application_libtelioapp_config_internalMeshnet_connectivityMatrix",
+            "set_context_application_libtelioapp_config_currentState_internalMeshnet_connectivityMatrix",
             Some(vec![val.as_str()]),
         ) {
             Ok(_) => Ok(Result::Success),